WIP
Added the modified script that needs a certain change to Eluna. The README inside the WIP folder has a link to the specific PR required, in case anyone wants to try it.

All it does at this time is teach a player all mounts bound to their account when obtaining the achievements for each riding skill, which eliminates the need to relog whenever they learn a new skill.